refactor(agent): extract shared utilities to shared.ts
